检索输出

+++ 08_DelegatesAndEventsSamples/ActionAndFunc/Calculator.cs
@@ -10,13 +10,13 @@ namespace ActionAndFunc
         {
             Console.WriteLine("I have 3 methods.");
         }
-        public double Add(int a, int b)
+        public int Add(int a, int b)
         {
             int result = a + b;
             return result;
         }
 
-        public double Sub(int a, int b)
+        public int Sub(int a, int b)
         {
             int result = a - b;
             return result;
